Bluetooth: L2CAP: Fix missing net_buf_unref()
In case of non-recoverable errors (e.g. the connection breaks while transmitting), the l2cap_chan_tx_process deques the net_buf but does not unreference it. As this is inside a work queue thread, the sending thread gets no information on this error, relying on the tx_process to ultimately free the buffer.

Signed-off-by: Patrick Rathje <git@patrickrathje.de>